requirements.txt
django
djangorestframework
psycopg2-binary
pillow
django-environ
gunicorn